+++ src/Donations/resources/admin/components/DonationDetailsPage/Tabs/Records/styles.module.scss
@@ -58,44 +58,6 @@
     gap: var(--givewp-spacing-4);
 }
 
-.statusSelect {
-    --statusOptionColor: var(--givewp-grey-500);
-    position: relative;
-
-    &::before {
-        background-color: var(--statusOptionColor);
-        border-radius: 50%;
-        content: '';
-        display: inline-block;
-        height: 0.75rem;
-        left: var(--givewp-spacing-4);
-        pointer-events: none;
-        position: absolute;
-        top: 50%;
-        transform: translateY(-50%);
-        width: 0.75rem;
-    }
-
-    &--pending,
-    &--processing {
-        --statusOptionColor: var(--givewp-blue-500);
-    }
-
-    &--publish {
-        --statusOptionColor: var(--givewp-green-500);
-    }
-
-    &--failed,
-    &--cancelled,
-    &--revoked {
-        --statusOptionColor: var(--givewp-red-500);
-    }
-
-    select.statusSelectInput {
-        padding-left: calc(var(--givewp-spacing-6) + 0.75rem);
-    }
-}
-
 .radioOptions {
     display: flex;
     gap: var(--givewp-spacing-6);
