+++ src/Subscriptions/resources/admin/components/SubscriptionDetailsPage/Tabs/Records/styles.module.scss
@@ -1,63 +1,5 @@
-.emptyState {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    gap: var(--givewp-spacing-4);
-    align-items: center;
-    text-align: center;
-    background: var(--givewp-shades-white);
-    padding: 40px 0;
-    height: 236px;
-}
 .description {
     font-size: 0.875rem;
     line-height: 1.43;
     color: var(--givewp-neutral-500);
-}
-
-.statusSelect {
-    --statusOptionColor: var(--givewp-grey-500);
-    position: relative;
-
-    &::before {
-        background-color: var(--statusOptionColor);
-        border-radius: 50%;
-        content: '';
-        display: inline-block;
-        height: 0.75rem;
-        left: var(--givewp-spacing-4);
-        pointer-events: none;
-        position: absolute;
-        top: 50%;
-        transform: translateY(-50%);
-        width: 0.75rem;
-    }
-
-    &--pending {
-        --statusOptionColor: var(--givewp-blue-500);
-    }
-
-    &--active {
-        --statusOptionColor: var(--givewp-green-500);
-    }
-
-    &--completed {
-        --statusOptionColor: var(--givewp-emerald-500);
-    }
-
-    &--expired,
-    &--cancelled,
-    &--failing,
-    &--trashed {
-        --statusOptionColor: var(--givewp-red-500);
-    }
-
-    select.statusSelectInput {
-        padding-left: calc(var(--givewp-spacing-6) + 0.75rem);
-    }
-}
-
-.notice {
-    margin-top: var(--givewp-spacing-1);
-    transform: translateX(1px);
 }
